# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-RECRUITMENTS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminStatsByStatus } from "./adminStatsByStatus";

export interface AdminStats {
  total_applications: number;
  total_jobs: number;
  total_users: number;
  total_organizations: number;
  by_status: AdminStatsByStatus;
}