# lib/api-zod/src/generated/types/createJobRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-RECRUITMENTS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateJobRequestEmploymentType } from "./createJobRequestEmploymentType";

export interface CreateJobRequest {
  title: string;
  description: string;
  organization_id: string;
  sector: string;
  location: string;
  deadline?: string;
  employment_type: CreateJobRequestEmploymentType;
}