# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-RECRUITMENTS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentType } from "./documentType";

export interface Document {
  id: string;
  user_id: string;
  name: string;
  type: DocumentType;
  url: string;
  size?: number | null;
  uploaded_at: string;
}