# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-RECRUITMENTS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: string;
  user_id: string;
  full_name?: string | null;
  email?: string | null;
  phone?: string | null;
  nationality?: string | null;
  date_of_birth?: string | null;
  address?: string | null;
  bio?: string | null;
  linkedin_url?: string | null;
  education?: string | null;
  experience?: string | null;
  skills?: string | null;
  languages?: string | null;
  is_complete: boolean;
  avatar_url?: string | null;
  created_at: string;
  updated_at: string;
}