# lib/api-zod/src/generated/types/submitApplicationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-RECRUITMENTS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitApplicationRequest {
  user_id: string;
  job_id: string;
  cover_letter?: string;
  applicant_name: string;
  applicant_email: string;
}