# lib/api-zod/src/generated/types/updateStatusRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-RECRUITMENTS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateStatusRequestStatus } from "./updateStatusRequestStatus";

export interface UpdateStatusRequest {
  status: UpdateStatusRequestStatus;
  applicant_email?: string;
  applicant_name?: string;
  job_title?: string;
}